backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Proton <feisuzhu@163.com>"]

[tool.poetry.dependencies]
python = "^3.8"
redis = "^3.4.1"
requests = "^2.23.0"
raven = "^6.10.0"
meinheld = "^1.0.1"
gunicorn = "^20.0.4"
django = "^3.0.3"
graphene = "^2.1.8"
django-annoying = "^0.10.6"
pillow = "^7.0.0"
graphene-django = "^2.8.2"
django-redis-cache = "2.1.1"
django-redis-sessions = "^0.6.1"
cryptography = "*"
itsdangerous = "^1.1.0"
msgpack = "^1.0.0"
pymysql = "^0.9.3"
psycopg2-binary = "^2.8.4"
pyyaml = "^5.3.1"
django-schema-graph = "^1.2.0"

[tool.poetry.dev-dependencies]
neovim = "^0.3.1"
jedi = "^0.17.2"
flake8 = "^3.8.4"
werkzeug = "^1.0.1"
pprintpp = "^0.4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
